Fix bottom-most chunk never meshing
The lowest chunk in a chunk column (at y0) was never meshed, because it couldn't find the below neighbouring chunk (in the void).

Fixed with a condition that "fakes" all chunks below the world to exist.
Added an assert that should trigger if infinite world height is added, as this "fix" would prevent infinite height.